Avoid race condition with opaque resource cache and Moodle session #581
The security check that tries to limit students to access resources
from the current question session is not working reliably, as Sarah Wood
found in Moodle 1.9, so take it out again.

Also, change the code to use slash-arguments, since it is nicer, and I
wrote the code thinking that was the problem we are having.