Fix a bug that causes duplicate rows on restore
Due to a bug in restore code, duplicate rows were created
in question extra data table (mdl_question_opaque). Because of
the fact that backup functionality is used to duplicate quizzes,
the number of rows in question_opaque table would double for each
quiz duplication operation that takes place in a related
course even if the quiz that is being duplicated does not include
any opaque questions.