Big reorganisation of the Opaque code.
This was the result of a careful review of all the code. Note that now
the code has been split between the question type and the behaviour in a
meaningful way. As part of that, there is a useful diagram
internaldoc/code_overview.dia/png. There is also a new test-case
internaldoc/testcase.txt.